Refractor models/submissions.py
Moves ApplicationRevision and AssignedReviewers model classes to the it's own file.

The submissions model file is quite big, this tries to make it little
lean.
