+++ hypha/apply/funds/models/mixins.py
@@ -136,7 +136,7 @@ class AccessFormData:
         try:
             return self.raw_data[definitive_id]
         except KeyError:
-            # We have most likely progressed application forms so the data isnt in form_data
+            # We have most likely progressed application forms so the data isn't in form_data
             return None
 
     @property
